requirements.txt
torch>=2.0.0
tensorflow>=2.12.0
numpy>=1.24.0
pandas>=1.5.0
scikit-learn>=1.3.0

# NLP
pymystem3>=0.2.2
tokenize-rt>=5.0.0

# Deep Learning
pytorch-lightning>=2.0.0
tensorboard>=2.13.0

# Data processing
pydantic>=2.0.0
pyyaml>=6.0

# Optimization & Scheduling
torch-optimizer>=0.13.0

# Metrics
scikit-metrics>=0.0.1

# Logging & Monitoring
clearml>=1.13.0
wandb>=0.15.0

# Utils
tqdm>=4.65.0
python-dotenv>=1.0.0
click>=8.1.0

# Development
pytest>=7.4.0
black>=23.0.0
isort>=5.12.0
flake8>=6.0.0
mypy>=1.5.0
